// src/models/matchaup.model.ts
export interface MatchupCreation {
    player1: string
    player2: string
    p1Score: number
    p2Score: number
    matchDate: string
    status: string
}

export interface Matchups {
    gameId: string
    player1: string
    player2: string
    p1Score: number
    p2Score: number
    matchDate: string
    status: string
}